requirements.txt
fastapi==0.95.0
uvicorn[standard]==0.24.0
sqlalchemy==2.0.20
alembic==1.17.2
psycopg[binary]==3.2.2
python-dotenv==1.1.1
passlib[argon2]==1.7.4
argon2-cffi==23.1.0
python-jose==3.3.0
pydantic==1.10.14
python-multipart==0.0.6
pytest==8.4.0
httpx==0.27.0
